apps/services/knowledge-base/requirements.runtime.txt
boto3==1.37.10
fastembed==0.7.4
fastapi==0.115.8
httpx==0.28.1
langchain-community==0.3.30
langchain-core==0.3.83
langchain-openai==0.3.35
langchain-qdrant==0.2.1
prometheus-client==0.21.1
psycopg[binary]==3.2.5
PyJWT==2.10.1
Pillow==11.1.0
pypdf==5.3.1
python-docx==1.1.2
python-multipart==0.0.20
qdrant-client==1.13.3
rapidocr-onnxruntime==1.4.4
uvicorn==0.34.0
